requirements.txt
whisperx==3.7.4
psycopg2-binary
requests
boto3
python-dotenv
